Query searcher and additional prefabs
A search box has been added that takes in a Cypher query. Does nothing at the moment except parse the string to be used in a query.
Also added more prefabs for the all the different node types in the mapador dataset.
Unfolding now also shows the right type of node.